Moved test_net, calc_logprob to lib/__init__.py
