+++ Chapter19/01_train_a2c.py
@@ -8,7 +8,7 @@ import roboschool
 import argparse
 from tensorboardX import SummaryWriter
 
-from lib import model, common
+from lib import model, common, test_net, calc_logprob
 
 import numpy as np
 import torch
@@ -26,31 +26,6 @@ ENTROPY_BETA = 1e-3
 ENVS_COUNT = 16
 
 TEST_ITERS = 100000
-
-
-def test_net(net, env, count=10, device="cpu"):
-    rewards = 0.0
-    steps = 0
-    for _ in range(count):
-        obs = env.reset()
-        while True:
-            obs_v = ptan.agent.float32_preprocessor([obs]).to(device)
-            mu_v = net(obs_v)[0]
-            action = mu_v.squeeze(dim=0).data.cpu().numpy()
-            action = np.clip(action, -1, 1)
-            if np.isscalar(action): action = [action]
-            obs, reward, done, _ = env.step(action)
-            rewards += reward
-            steps += 1
-            if done:
-                break
-    return rewards / count, steps / count
-
-
-def calc_logprob(mu_v, logstd_v, actions_v):
-    p1 = - ((mu_v - actions_v) ** 2) / (2*torch.exp(logstd_v).clamp(min=1e-3))
-    p2 = - torch.log(torch.sqrt(2 * math.pi * torch.exp(logstd_v)))
-    return p1 + p2
 
 
 if __name__ == "__main__":
