+++ Chapter19/04_train_ppo.py
@@ -8,7 +8,7 @@ import roboschool
 import argparse
 from tensorboardX import SummaryWriter
 
-from lib import model
+from lib import model, test_net, calc_logprob
 
 import numpy as np
 import torch
@@ -29,33 +29,6 @@ PPO_EPOCHES = 10
 PPO_BATCH_SIZE = 64
 
 TEST_ITERS = 100000
-
-
-def test_net(net, env, count=10, device="cpu"):
-    rewards = 0.0
-    steps = 0
-    for _ in range(count):
-        obs = env.reset()
-        while True:
-            obs_v = ptan.agent.float32_preprocessor([obs]).to(device)
-            mu_v = net(obs_v)[0]
-            action = mu_v.squeeze(dim=0).data.cpu().numpy()
-            action = np.clip(action, -1, 1)
-            if np.isscalar(action): action = [action]
-            if np.isscalar(action): action = [action]
-            obs, reward, done, _ = env.step(action)
-            rewards += reward
-            steps += 1
-            if done:
-                break
-    return rewards / count, steps / count
-
-
-def calc_logprob(mu_v, logstd_v, actions_v):
-    p1 = - ((mu_v - actions_v) ** 2) / (2*torch.exp(logstd_v).clamp(min=1e-3))
-    p2 = - torch.log(torch.sqrt(2 * math.pi * torch.exp(logstd_v)))
-    return p1 + p2
-
 
 def calc_adv_ref(trajectory, net_crt, states_v, device="cpu"):
     """
