+++ Chapter19/06_train_sac.py
@@ -9,7 +9,7 @@ import argparse
 from tensorboardX import SummaryWriter
 import numpy as np
 
-from lib import model, common
+from lib import model, common, test_net
 
 import torch
 import torch.optim as optim
@@ -27,26 +27,6 @@ REPLAY_INITIAL = 10000
 SAC_ENTROPY_ALPHA = 0.1
 
 TEST_ITERS = 10000
-
-
-def test_net(net, env, count=10, device="cpu"):
-    rewards = 0.0
-    steps = 0
-    for _ in range(count):
-        obs = env.reset()
-        while True:
-            obs_v = ptan.agent.float32_preprocessor([obs]).to(device)
-            mu_v = net(obs_v)
-            action = mu_v.squeeze(dim=0).data.cpu().numpy()
-            action = np.clip(action, -1, 1)
-            if np.isscalar(action): action = [action]
-            obs, reward, done, _ = env.step(action)
-            rewards += reward
-            steps += 1
-            if done:
-                break
-    return rewards / count, steps / count
-
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
